requirements.txt
Flask==2.3.2
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.10.0
schedule==1.1.0
pandas
mysql-connector-python
openpyxl